Update `Topic.ParentID` when a topic is moved
Previously, when a topic was moved, the `ParentID` was being updated in the `Attribute` table. With the migration of core attributes to the `Topics` table, this is no longer necessary—and, worse, leaves the core `Topics` record in an invalid state (with the `ParentID` pointing to a different parent than the hierarchy represents). This is fixed by updating the logic in `MoveTopic` to correctly update the `Topics` table.