Remove `MarkClean()` during `TopicRepositoryBase.Move()`
This doesn't _hurt_ anything, but it also doesn't _do_ anything since the `ParentId` is no longer stored as an attribute and, therefore, the `Attributes` collection should no longer be marked as `IsDirty()` do to the `ParentId` having been modified.

(There's an argument that this _should_ mark the main `Topic` as clean. Currently, the main topic doesn't get marked as `IsDirty()` when the `Parent` changes. We may need to revisit that later.)