Ensured `NULL` attribute values are converted to empty values
The `AttributeValue` column is not nullable, and `NULL` values (e.g., to overwrite previous versions) are expected to be submitted as empty strings. To make the interface fool-proof, this is enforced within the `UpdateAttributes` stored procedure, thus allowing `NULL` attributes to be submitted, but ensuring they're properly converted to an empty string.