Ensured `GetChildTopicIDs` can accept a `NULL` parent
The `ParentID` can be null in the `Topics` table—though this is usually needed exclusively for the `Root` node. Previously, this wasn't accounted for in the `GetChildTopicIDs` function, thus preventing it from e.g. retrieving the root node by including a `@ParentID` parameter of `NULL`. This is fixed in this update.