Ensured `FindTopicIDs` can accept a `@AttributeValue` of `NULL`
Most attribute value columns are not nullable. As such, if a `NULL` value is submitted, it won't match anything. That's fine, given the nature of the procedure. But, normally, we'd expect it to return empty values in that case—e.g., attributes that were overwritten. (These are always treated as empty.)

Further, for the core attributes, the `ParentID` column _can_ be `NULL`. As such, this _also_ ensures that `FindTopicIDs` can find, if appropriate, the root node, similar to a previous update to `GetChildTopicIDs` (9f42085).