Ensured transaction is completed on error
Previously, any errors would result in an immediate return. This left the transaction incomplete, however. To fix this, the transaction—in which nothing has actually been done—will be immediately committed prior to returning. That resolves a bug in which case the following call to `MoveTopic` would fail due to the previous transaction being incomplete.