Renamed and improved `TopicRenameEventArgs`
Renamed `RenameEventArgs` to `TopicRenameEventArgs` to better clarify that it's intended to map to topic operations; this is consistent with .NET CLR conventions. Added the `OriginalKey` and `NewKey` parameters, properties to fully map the range of options available for rename events. Technically, this information can be inferred from the `Topic`, but this makes it more explicit.