Updated SQL objects to use high-precision `DATETIME2`
Microsoft recommends using `DATETIME2` over `DATETIME` in SQL because it has a broader range and more flexibility over precision. By using a precision of (7), we get the same precision as C#'s `DateTime` object, and thus avoid truncation issues which have caused problems in the past. This allows us to set the version with confidence in C# knowing the return value will be the same. It also allows us to not worry as much during unit tests or automated updates about version conflicts occuring due to overlapping timestamps.

(In practice, we probably don't need versions to be _this_ precise, and we _may_ reevaluate that later. But we should definitely be using `DATETIME2`, and can revisit the precision in a future update if necessary.)