Use integer for `NULL`, not empty
Since `dbo.Topics.ParentID` is nullable, we wrap it in an `ISNULL()` check. Previously, out of habit, I defaulted to returning a string for empty. This required doing something similar for `@TopicID`. But `@TopicID` is an `INT`, so converting from an empty string isn't explicit or predictable. Instead, preferably, we'll convert both to an explicit `INT` that's outside the expected range of the table.