Allow `AttributeDescriptor` to be initialized directly
Previously, `AttributeDescriptor` was marked as `public`, but we moved it to `abstract` in OnTopic 4.0.0 in preference for strongly typed `AttributeDescriptor` classes, such as `TextAttribute`. In OnTopic 5.0.0, we're moving those strongly typed `AttributeDescriptor` classes into their corresponding OnTopic Editor plugins (1d122e0). That allows `AttributeTypeDescriptor` instances to be defined dynamically by including external packages.

As a result, however, that means that these types won't be locally available unless that plugin is included. If an application doesn't expose the OnTopic Editor, there isn't any reason to include that plugin. In that case, these attribute classes will fall back to being instantiated as a `Topic`. That prevents them from being used as `AttributeDescriptor`s during e.g. `SqlTopicRepository.Save()`.

But, outside of OnTopic Editor, the properties most applications need to read for attribute descriptors are on the main `AttributeDescriptor` type itself. As such, we're making it `public` again so that it can be initialized as a fallback for cases where the strongly typed versions aren't in scope.

A subsequent update will include this logic in the `DynamicTopicLookupService`.