Added `AttributeDescriptor` fallback logic for `DynamicTopicLookupService`
As discussed in a previous commit (4116633), the strongly typed versions of `AttributeDescriptor` are now contained within plugins for the OnTopic Editor (1d122e0), and will only be within scope if the OnTopic Editor is configured and that plugin is included.

In absence of that, however, most other applications rely on members provided by the more general `AttributeDescriptor`. Given that, instead of falling back to the default of `Topic`, we want to fall back to `AttributeDescriptor` if the requested `typeName` ends in `AttributeDescriptor`, but the specialized type cannot be found.

This ensures that e.g. the `ContentTypeCollection.AttributeTypes` collection is properly initialized, and that code that needs general access to attribute type configuration, such as `SqlTopicRepository`, can still get access to that, even if they don't have the more precise overrides from e.g. `EditorType` or `ModelType` typically exposed by the subclasses.

This allows the attribute types to be decoupled from the OnTopic library—as well as the OnTopic Editor—and thus enabled plugins to be dynamically included without introducing problems for applications where those plugins aren't in scope.