Established new `ObservableTopicRepository` base class
The `TopicRepositoryBase` is a very opinionated base class. That's useful! But there are cases where that level of opinion isn't necessary or even desirable, such as the `CachedTopicRepository`. The `ObservableTopicRepository` separates out the basic event handling logic, which is expected to be shared between all implementations—and isn't usually expected to vary significantly—into its own base class.

In a future update, the `TopicRepositoryBase` will be updated to take advantage of this, as will decorators, such as the `CachedTopicRepository`, which don't otherwise need the overhead of `TopicRepositoryBase`.