Established new `TopicRepositoryDecorator` base class
The `ITopicRepository` is a prime candidate for decorators. Decorators require very specific handling to register the underlying implementation and wire-up passthroughs for both events and members. Outside of that, decorators don't require much in terms of implementation, and thus don't benefit from deriving from `TopicRepositoryBase`.

The `TopicRepositoryDecorator` mitigates this by offering a base class just for decorators. It handls all of the wiring so that implementors don't need to worry about that, and can focus exclusively on the parts they need to override.