Renamed `TopicRepositoryBase` to `TopicRepository`
The .NET Framework Design Guidelines recommend against using the `Base` suffix for publicly available classes. For example, the abstract `Controller` class is named `Controller`, not `ControllerBase`. This should be applied to `TopicRepositoryBase` as well. The fact that it doesn't specify a persistence store (e.g., `SqlTopicRepository`) or any functionality (e.g., `CachedTopicRepository`) should be a clear enough hint that it's not intended for direct usage, since `ITopicRepository` implementations should offer specific implementation logic.

While I was at it, I improved the documentation for `TopicRepository` to better articulate when it should be used, and what alternatives are available.