Centralized core business logic from `SqlTopicRepository.Save()`
The `SqlTopicRepository.Save()` implementation had a lot of logic that wasn't specific to SQL Server, and would likely need to be repeated by subsequent implementations. To help avoid that, it has been migrated from `SqlTopicRepository.Save()` and into the underlying `TopicRepository.Save()`.

This includes tracking `unresolvedTopics` (i.e., topics whose relationships or references haven't yet been saved) and then resaving them at the end. It also includes the core recursion logic for looping over child topics, if `isRecursive` is set. Finally, it includes establishing a shared `version` to use across all attributes, relationships, and references so all updates that are part of this operation share the same version.

While not yet implemented, this will also allow us to extend the unit testing capabilities to evaluate the above features, as the unit tests don't currently cover `SqlTopicRepository.Save()`, but do cover `TopicRepository.Save()`.

As part of this, the `Save()` method is marked as `sealed` and the core implementation is moved to the `abstract` `Save(topic, version, persistRelationships)` overload. This prevents the need for derived implementations to call `base.Save()`, and likewise allows the `TopicRepository.Save()` implementation more control over the order of operations—e.g., by placing validation logic above the implementation, while putting topic graph operations after it.

This fixes a behavior where an operation would be reflected in the local topic graph even if the database operation failed, which often resulted in confusion since it would appear as though the operation completed based on the cached state of the topic graph. This could also potentially result in bugs in subsequent saves, since those operations would have already been completed.

Finally, this includes updates to the `StubTopicRepository` to `override` the updated `Save()` overload, instead of the `ITopicRepository` version.