Delegated core implementation to `{Operation}Topic()` methods
Previously, derived implementations were expected to override the main `ITopicRepository` methods, such as `Save()`, `Move()`, and `Delete()`. That necessitated that they call the `base.{Operation}()` as part of that call, while simultaneously requiring the `TopicRepository` implementation to handle all business logic in one sequence—regardless of made more sense to run it before or after the actual business logic. In practice, implementors were expected to place the `base{Operation}()` call before the implementation logic, but there was no way of strictly enforcing this.

These issues are mitigated by delegating the core implementation to new `abstract` `{Operation}Topic()` methods—such as `SaveTopic()`, `MoveTopic()`, and `DeleteTopic()`. These have no need to call the `base.{Operation}()` method.

As part of this, the core `{Operation}()` methods are now marked as `sealed` so they can't be overriden directly. That ensures that their business logic is always run. It also allows them to call the `{Operation}Topic()` implementations in the most appropriate place—i.e., after validating the parameters and conditions, but before persisting the values to the topic graph or raising the event.

This fixes a behavior where an operation would be reflected in the local topic graph even if the database operation failed, which often resulted in confusion since it would appear as though the operation completed based on the cached state of the topic graph. This could also potentially result in bugs in subsequent saves, since those operations would have already been completed.