Remove legacy `DerivedTopic` hack
Previously, the `DerivedTopic` needed to be reset to itself to force the `DerivedTopic` (then `TopicID`) attribute to be resaved to the `Topic.Attributes` collection so that updates to its `Topic.Id` from the `Save()` operation could be updated. Without this, if the `DerivedTopic` hadn't yet been saved, then the original `Save()` operation would have written a `TopicID` of `-1` to the database, and never have been updated with the final post-save `Topic.Id`.

As the `DerivedTopic` is now handled as part of the `Topic.References`, this is all handled in a more general fashion, via the `unresolvedTopics` logic, without the need for a specific hack to handle `DerivedTopic`. In fact, based on how `Topic.References` work, this wouldn't do anything even if there was alternate logic, since the `Topic.References` operates off references to the target `Topic`, not a `TopicID` attribute.