Disable `IsFullyLoaded` if there's an exception on `Load()`
If an exception occurs during `Load(topicId, version, referenceTopic)`, we don't have confidence that the topic's relationships or references were properly loaded. They _might_ have been, but it's not possible to tell for sure. As such, the target `topic` may be left in an incomplete state. To mitigate this from potentially resulting in data loss during a subsequent `Save()` operation, these collections are explicitly marked as `!IsFullyLoaded`. This is especially important since those collections are cleared prior to the operation, since it's otherwise impractical to track relationships and references that were added after the loaded `version`.

This is hopefully a rare scenario, but it will help avoid inadvertent data loss in the process—though it may also result in deleted relationships and references not being deleted. (The OnTopic Editor will warn editors about this possibiliy before they save a topic in this state.)