Throw exception during `TopicRepository.Save()` if `unresolvedTopics`
Something, unresolved relationships or topic references will occur during a `Save()`. This happens if the topics that are references haven't yet been saved themselves; in that case, they can't be persisted to the database. That's no problem; the `TopicRepository.Save()` keeps track of those, and automatically attempts to resave them after the `Save()` operation has been completed.

If, after that operation is complete, there are _still_ `unresolvedTopics`, however, then a `ReferentialIntegrityException` will be thrown. This usually occurs if the relationships or topic references point to newly created topics outside the scope of the `Save()`, and can usually be fixed by expanding the scope of the `Save()` or first saving those referenced topics.