Updated new `ITopicRepository` classes to use renamed event arguments
Previously, the improvement/ITopicRepository-events branch renamed the `DeleteEvent`, `MoveEvent`, and `RenameEvent` to `TopicEvent`, `TopicMoveEvent`, and `TopicRenameEvent`. In the meanwhile, in the improvement/TopicRepositoryBase-refactoring branch, additional base classes implementing `ITopicRepository` were introduced which referenced these classes using the legacy names. These needed to be updated to use the latest naming conventions.