Removed `virtual` from the `RedirectController.Redirect()` action
This implementation is so straight forward that it's likely anyone with specialized needs would just create their own controller. This is really just a convenience controller for a standard requirement; customers with more sophisticated needs have added their own redirect controllers either to replace or to complement the out-of-the-box `RedirectController`.

If we discover a specific extensibility point that we want to customize in the future, we can reevaluate the best approach for that then.