Removed `virtual` from the `SitemapController` actions
In the `SitemapController`, the implementation is so involved that it's unclear how deriving this would satisfy any customization requirements.

If there's a need to customize the business logic—and there may well be!—we'd want to revisit what the extensibility points are, and whether there's a better way of accomodating the need. For instance, there may be a good argument for exposing an options object to the constructor, or a delegate which validates which topics or attributes to include in the index.

For now, however, having the `Index` and `Extended` actions marked as `virtual` doesn't really facilitate those potential requirements.