Removed `virtual` from `TopicView()` helper
It's unclear exactly how a derived class might expect to update this. If they needed to customize `TopicViewResult` then they'd likely want a new method, and would likewise expect extensive updates to the underlying infrastructure components that this feeds into down the pipeline. The purpose of this is really just to tie in the contextual `ViewData`, `TempData`, and `ContentType` so it doesn't need to be explicitly defined.