Removed `virtual` from the event definitions
Originally, these were marked as `virtual` so that they could be overwritten in decorators, such as the `CachedTopicRepository`, in order to handle event bubbling.

With the introduction of the raise methods (e.g., `OnTopicDeleted()`), this can be handled in the constructor by simply subscribing to the source `ITopicRepository` events and relaying them to the raise methods.

Given that, there's no strict need to keep these marked as virtual. If we rediscover a need to do this, we can revisit that in the future.

(Technically, since the only reason we moved away from the field-like event syntax in the first place was to allow them to be `virtual`, we could go back to the symplified event format. But since this is now in its own `ObservableTopicRepository` base class that's not especially critical.)