Removed `virtual` from `SetContentTypeDescriptors()` method
There aren't currently any implementations that override these and it's unclear what specific need they'd be addressing if they did. As always, we can revisit this down the road if there's a customer requirement. But, even if we did, we'd likely want to revisit how this was handled—such as, potentially, a new `protected virtual` overload with no implementation which allows business logic to be _injected_ into the main overload, instead of relying on the derived class to call `base.SetContentTypeDescriptors()`.