Mark helper methods as `private` in mapping services
In the `TopicMappingService` and `ReverseTopicMappingService`, we had a number of helper methods—such as `SetScalarValue()` or `GetSourceCollection()`—which were marked as `protected`. Presumably, this was so that future derived classes could reuse these, while overriding other aspects of the functionality. In practice, though, none of the members are marked as `virtual`, and so there's no (supported) use case for extensibility.

If we want to allow the `TopicMappingService` or `ReverseTopicMappingService` to be extended, we should carefully evaluate what the use cases are, and better address them through e.g. `protected virtual` members, delegates, and possibly an options object. Until then, we should not expose internal helper methods with a more deliberate plan for how they will be used, and care to make sure we're protecting core functionality while exposing room for expansion.

Finally, the `BindingModelValidator`, which is related to the `ReverseTopicMappingService`, already had its protected members migrated to `internal` (not `private`, since they're used by `ReverseTopicMappingService`), but the comment headers still communicated that they were `protected`, so that's been resolved while I was at it.