Removed validation of boolean types in `TopicEventArgs` derivatives
The nullability analysis doesn't necessitate that we validate whether non-nullable value types are null. More importantly, when using `Contract.Requires()` with a Boolean value, the semmantics are confused, as this is checking to determine if a condition is true. As such, `Contract.Requires()` will actually fail if the parameter is set, but set to `false`. That's not what we want. Whoops!