Removed check for `TopicId` in `GetValue()` fallback
It's not entirely clear why this condition was here, as it's redundant with the next condition, but a check for the legacy `TopicId` attribute will no longer work going forward, as that value is now represented by the topic reference with the key `DerivedTopic`. Regardless, the correct way to evaluate that remains to check if `DerivedTopic` is `null`.