Added `CompressHierarchy` as part of post-test cleanup
Normally, we don't need to run `CompressHierarchy` on the topics hierarchy. The unit tests, however, take a few shortcuts when bulk deleting topics—such as deleting all records with a key starting with the test name—instead of calling `DeleteTopic`, which properly closes any gaps. Given that, we need to run `CompressHierarchy` at the end to ensure that there aren't large gaps left over after running the unit tests.

(These gaps don't hurt anything, though they make it harder to debug the hierarchy if there's a need—and, over a long period, could hypothetically overflow the maximum range of the `INT` field if left unchecked.)