Allow `ParentID` to be `NULL` for `CreateTopic`, `MoveTopic`
This will optionally allow new topics to be created in the root. This isn't normally recommended, but currently it results in a malformed hierarchy, so we want to make sure that doesn't happen. (We do create topics in the root as part of the unit tests, in order to minimize variables.)