Correctly set `@InsertionPoint` in `MoveTopic` if `@ParentID` is `NULL`
Previously, if the `@ParentID` was `NULL` (or `-1`), the `@InsertionPoint` wasn't defined, and thus would end up getting inserted in an overlapping range with the root node. This addresses that, by placing it at the end of the root node(s).

(This change looks a bit more invasive than it is because I rearranged the order of the logic to keep the conditions simple.)