Added missing `Slideshow` view
Slideshow is one of the built-in content types exposed by one of the out-of-the-box view models. There should be a corresponding view in the `Host` test site.

Note: Technically a slideshow is a type of `ContentList` and could just share that. I've duplicated the code, however, since `Slideshow` also exposes a custom `TransitionEffect` attribute.