Introduced specialized `DirtyKeyCollection`
The `DirtyKeyCollection` is an `internal` collection meant to help standardize how dirty keys are tracked, by moving that logic out of disparate collections and instead into a single utility collection. It implements the new `ITrackDirtyKeys` interface, but also introduces a couple of more specialized methods, such as `MarkDirty()` and `MarkAs()`.