Implemented `DirtyKeyCollection` in `TopicReferenceDictionary`
This (slightly) simplifies the (admittedly simple) logic in `DirtyKeyCollection` by centralizing the implementation of these methods. As the `TopicReferenceDictionary` didn't previously track individual keys, this actually extends the functionality slightly, and makes it marginally more intelligent. For instance, if an item is removed from an unsaved topic, then it marks the key as clean, not dirty, since it knows it cannot have been previously saved.