Implemented `DirtyKeyCollection` in `TopicRelationshipMultiMap`
This (slightly) simplifies the (admittedly simple) logic in `DirtyKeyCollection` by centralizing the implementation of these methods. This also improves the `IsDirty()` handling of `TopicRelationshipsMultiMap` by allowing it to mark a key as clean when it's being removed if it points to an unsaved `Topic`; in that case, we know that it hasn't previously been saved, and thus should not mark the collection as `IsDirty()`.