Applied `ITrackDirtyKeys` to `TopicReferenceDictionary`
This required introducing some overloads that hadn't previously been accounted for—namely `IsDirty(key)` and `MarkClean(key)`—since `TopicReferenceDictionary` hadn't previously track individual keys before the introduction of `DirtyKeyCollection` (9705c15).

As a result, these overloads introduced ambiguity in some of the XML Doc references in teh `TopicReferenceDictionaryTest`, so they needed to be updated to refer to the specific overload.