Updated unit tests to account for new `Remove()`, `Clear()` logic
With the introduction of the new `DirtyKeyCollection`, the `TopicReferenceDictionary` and `TopicRelationshipMultiMap` now use more intelligent logic on `Remove()` and `Clear()`, only marking the key as dirty if the current (source) topic hasn't been saved, and otherwise marking it as clean.

To account for this, the `TopicReferenceDictionary` and `TopicRelationshipMultiMap` tests needed to be updated to assign an `Id` to the test topics so that the `IsDirty()` state can correctly be set.