Fixed bug in `RemoveTopic()`'s `MarkDirty()` handling
In a previous update, I had introduced an optimization to mark the `relationshipKey` as clean if the target `topic` was not saved. That makes sense in that we know that a `topic` that's `IsNew` cannot have been saved previously, and thus shouldn't affect the `IsDirty()` state. But that doesn't account for _other_ topics in the relationship. Removing a new topic from the collection shouldn't mark the collection as clean if there are _other_ topics in the collection that _may_ have been marked as dirty.

This could be accounted for by adding a check to see if _all_ items in the collection are `IsNew`, in which case this logic would be appropriate. But this is such a corner case that the additional complexity and overhead of a more sophisticated check doesn't really make sense here.