Removed unused overloads for `ReferentialIntegrityException`
The overload that accepts a `sourceTopic` is hard-coded for handling missing or unsaved `DerivedTopic` references. In practice, referential integrity exceptions can be thrown in a variety of circumstances, and the `DerivedTopic` case is more generally handled by validating the new `Topic.References` when saving a topic.

Given that, these overloads are not used. Not only that, but while it might be tempted to attempt to generalize the messaging so that a `sourceTopic` and a `targetTopic` can be passed in, even that introduces challenges since a referential integrity exception _could_ potentially mean that one of those values cannot be found.

We may revisit this later if we identify a use case for it, but for now this scenario is being removed.