Updated to 2.0.0 release of `Microsoft.Data.SqlClient`
This is one of the few production dependencies of the OnTopic Library which is relayed to implementers; the rest are mostly build dependencies. Since this is a public library, we try to keep production dependencies on their lowest possible version in order to maximize backward compatibility.

That said, as this is a major release that is over six months old, and this is a major update to OnTopic, we think it's reasonable to expect implementors to upgrade to the latest major version of `SqlClient` while they're upgrading to OnTopic 5.0.0.

Still, we're only requiring `SqlClient` 2.0.0. While we expect customers will likely want to use the latest version, we're allowing flexibility in which version they adopt, in case there are e.g. particular bugs or limitations introduced by minor updates or patches, and which affect other operations outside of OnTopic.