Consolidated overloads of `TopicFactory.Create()`
There were two overloads of `TopicFactory.Create()`, which can be satisfied with one overload with optional parameters assuming a different parameter ordering. That paramter ordering is actually consistent with the respective parameter ordering of the `Topic` constructor itself, and thus makes more sense for consistency anyway. As such, I've consolidated these two overloads into one, and flipped the order of the (optional) `parent` and `id` parameters.

Technically, this is a breaking change. But, generally, the only codes that call the full `TopicFactory.Create()` overload are the `ITopicRepository` implementations—namely `SqlTopicRepository`—and the unit tests. Outside of that, most callers of `TopicFactory.Create()` will only be passing the `key`, `contentType`, and `parent` parameters, which will remain compatible with this ordering.