Introduced `TopicFactory.Create()` overload with `id`, but not `parent`
It's really rare that production code will need to create a `Topic` with an `id` but not a `parent`. But this use case happens frequently in test cases, where we want to simulate a saved entity (by assigning it an `Id`) but don't want to make it part of the broader topic graph.

Technically, this could probably be marked `internal` since it is expected to primarily be used by unit tests. But we'll maintain it as it helps provide backward compatibility with the previous overloads, prior to merging the two main `Create()` implementations (99d6d51).