Updated unit tests to account for `TopicFactory.Create()` parameter order
In a previous update, I consolidated two overloads of `TopicFactory.Create()` (99d6d51), which resulted in changing the parameter order of the last two optional parameters (`parent`, and `id`). This order is actually consistent with the `Topic` constructor and, thus, preferred.

It's really rare for code to provide all of these parameters and, thus, it normally doesn't matter. The main use case for this is when loading data from persistence stores, and so the only production code we'd expect to call it is the `ITopicRepository` implementations—and, specifically, the `SqlTopicRepository` (as the rest are all test doubles, base classes, or decorators).

That said, the one _other_ place where we use all four parameter are the unit tests, as they're needed when a) defining a topic graph, and b) assigning each entity a fake identifier (to simulate a saved state). Given that, the unit tests—and the `StubTopicRepository` they rely on—need to be updated to reflect the new order.