Updated `ITypeLookupService.Lookup()` to accept a `params string[]`
Previously, the `ITypeLookupService`'s `Lookup()` method required a single `string typeName` parameter. Now, that has been updated to a `params string[] typeNames` parameter. This allows a prioritize list of fallbacks to be supplied in order to support multiple acepted naming conventions and/or acceptable defaults. The first match will be returned.

This will break all `ITypeLookupServices`. The following commits will resolve these issues in the OnTopic library. External libraries, such as the OnTopic Editor, as well as third-party implementations will need to be updated to reflect this change.