Removed the `StaticTypeLookupService.DefaultType`
Previously, the `StaticTypeLookupService` offered a `DefaultType` property which allowed each implementation to have a default fallback established that would be returned if the supplied `typeName` isn't found.

In practice, this often led to confusion, since implementations didn't know what that `DefaultType` was, and had no way to access it off the `ITypeLookupService` abstraction and, therefore, had to evaluate the result against their input to determine if a concrete match was returned or not.

Since we can now specify alternative or fallback types via the `Lookup()` method, there's no need for the `DefaultType` anymore. Instead, the preferene is for callers to specify a default type that's acceptable to them as the last argument in their `typeNames`. This provides more transparency, flexibility, and control to the callers, and helps ensure that there aren't unexpected consequences when swapping out `ITypeLookupService` implementations.

This breaks all derived classes. Those that are part of the OnTopic library will be updated in subsequent commits. External libraries, such as the OnTopic Editor, and third-party implementations may need to update any `ITypeLookupService` implementations they have to reflect this change in the constructor.