Updated `StaticTypeLookupService` derivatives `base()` constructor
With the change of the `StaticTypeLookupService`'s constructor to no longer accept an optional `defaultType` parameter, each derivative implementation must be updated to remove this `defaultType`. If fallbacks are necessary, they will need to be accounted for in calling code—which will be implemented for the OnTopic Library in subsequent updates.