Moved fallback logic for `TopicLookupService`s to `TopicFactory`
With the preference against relying on the `ITypeLookupService` implementations to handle fallback and defaults, the `TopicFactory` has been updated to handle the `AttributeDescriptor` alternative and `Topic` default, instead of relying on the `DefaultTopicLookupService` or the `DynamicTopicLookupService` to handle this.

Positively, this centralizes the `AttributeDescriptor` fallback logic, which was previously repeated in the `DefaultTopicLookupService` as well as the `DynamicTopicLookupService`, since they didn't have a common topic-specific base class.

Technically, this implementation _could_ have utilized the new `Lookup(params)` argument by specifying these defaults in the arguments list. Since they're well-known types, however, and are never expected to be replaced, the code is a bit more readable in this format, which consolidates the fallback with the necessary null check.