Moved fallback logic for `TopicViewModel`s to `TopicMappingService`
With the preference against relying on the `ITypeLookupService` implementations to handle fallback and defaults, the `TopicMappingService` has been updated to handle the `ViewModel` alternative, instead of relying on the `DynamicTopicViewModelLookupService` or the `TopicViewModelLookupService` to handle this.

Positively, this centralizes the `ViewModel` fallback logic, which was previously repeated in the two implementations, since they didn't have a common view model specific base class.

This also allows us to simplify the detection of a miss, since we don't need to contend with the expectation of a default value being returned. (Technically, an `ITypeLookupService` could still opt to do this, but that is recommended against.)