Updated `CompositeTypeLookupService` to loop through `typeNames`
The `CompositeTypeLookupService` needs to use care when working with the `params string[] typeNames` array. A naive implementation might simply relay the `typeNames` to the `Lookup()` method of each underlying `ITypeLookupService` implementation. That would result in prioritizing the fallbacks, however, in the case that the first implementation couldn't find a prioritized match.

Instead, it must loop through each `typeName` and pass it to each `Lookup()` method, thus ensuring that the priority of both the `typeNames` as well as the `ITypeLookupService` implementations is honored.