Evaluate fallbacks using new `params string[] typeNames`
Previously, fallbacks were evaluated using built-in fallback logic in the concrete `ITypeLookupService` implementation. This is a confusing approach and prone to introducing not only repetitive logic (between implementations) but also errors when swapping out implementations.

To mitigate that, this is now controlled specifically by the caller when doing a `Lookup()`, by specifying any acceptable alternatives or defaults as part of the `params` list.

These unit tests are updated to reflect that.