Renamed `DerivedTopic` to `BaseTopic`
The name `DerivedTopic` doesn't make semantic sense. It points to the `Topic` that the current `Topic` is derived _from_, but the name suggests the opposite—that it refers to a `Topic` that that derives from the _current_ `Topic`.

This was previously addressed in the OnTopic Editor by labeling this attribute as "Inherited Topic". But even that is a bit ambiguous, and used sometimes to refer to the base class and other times to refer to the derived class. In the meanwhile, the code has continued to use `TopicDescriptor` for backward compatibility.

With so many breaking changes queued up for the OnTopic 5.0.0 release, this is a good opportunity to finally resolve this nomenclature. While, technically, `InheritedTopic` should be accurate, the inconsistency of its use doesn't make it a marked improvement over `DerivedTopic`. Given that, we're going with `BaseTopic`—which isn't ideal, but is semantically unambiguous.

(**Note:** Microsoft often uses a property called "Base" for these circumstances. That itself is a bit confusing since the `base` keyword refers to the superclass, but the `Base` class refers to an instance whose values are inherited, not a class whose members are inherited. The purpose of `BaseTopic` is to help clarify this.)

This will almost certainly break downstream libraries and implementations. This is particularly true of the OnTopic Editor, which has special handling for the inherited topic. This also breaks the actual database implementation, and necessitates that existing topic references have their names updated; that will be handled via the database migration script in a subsequent commit.